% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amino_acid_codes}
\alias{amino_acid_codes}
\title{String of valid amino acid codes}
\format{An object of class \code{character} of length 1.}
\usage{
amino_acid_codes
}
\description{
String of valid amino acid codes
}
\keyword{datasets}
