% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageMSD.R
\name{averageMSD}
\alias{averageMSD}
\title{Generate an average of three MALDI spectrum samples}
\usage{
averageMSD(path, pdr, resolution = 0.01)
}
\arguments{
\item{reps}{a data structure holding three objects, S1, S2 and S3, each of which has a column called `mass` and `intensity`}

\item{stats}{approx ccf ksmooth lm predict}
}
\value{
a dataframe holding the following fields:
  \item{mass}{The interpolated mass values}
  \item{intensity}{The average intensity values}
}
\description{
Generate an average of three MALDI spectrum samples
}
