% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_gxy_probs.R
\name{ba_gxy_probs}
\alias{ba_gxy_probs}
\title{Calculate the 'GXY' frame of a peptide, and use probabilities of hdroxylation for prolines at different positions
to estimate the probability of every possible level of hydroxylation}
\usage{
ba_gxy_probs(seq, xprob = 0.05, yprob = 0.95, xyxprob = 0.5,
  xyyprob = 0.99, verbose = F)
}
\arguments{
\item{seq}{a peptide sequence}

\item{xprob}{the probability of hydroxylation at the X position}

\item{yprob}{the probability of hydroxylation at the Y position}

\item{xyxprob}{the probability of hydroxylation at the X position if the Y position is also Proline}

\item{xyyprob}{the probability of hydroxylaiton at the Y position if the X position is also Proline}

\item{verbose}{set to TRUE for more verbose processing}
}
\description{
Calculate the 'GXY' frame of a peptide, and use probabilities of hdroxylation for prolines at different positions
to estimate the probability of every possible level of hydroxylation
}
\examples{
ba_gxy_probs
}
\keyword{"Mass}
\keyword{spectrum",}
