% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba_plotseqpeaks.R
\name{ba_plotseqpeaks}
\alias{ba_plotseqpeaks}
\title{plot the peaks of a mass spec}
\usage{
ba_plotseqpeaks(data, myxlim)
}
\arguments{
\item{data}{a data strucyture in MALDI format}

\item{myxlim}{c(min,max) for the range of the data you want to plot}
}
\description{
plot the peaks of a mass spec
}
\keyword{bruker}
