% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioarch_mammal_sequences}
\alias{bioarch_mammal_sequences}
\title{Collagen sequences for a range of mammals generated by M.Collins / BioArch}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 2102 columns.}
\usage{
bioarch_mammal_sequences
}
\description{
Collagen sequences for a range of mammals generated by M.Collins / BioArch
}
\keyword{datasets}
