% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bms0.1.2}
\alias{bms0.1.2}
\title{EXPERIMENTAL "versioned" copy of bioarch_mammal_sequences}
\format{An object of class \code{data.frame} with 11 rows and 2102 columns.}
\usage{
bms0.1.2
}
\description{
EXPERIMENTAL "versioned" copy of bioarch_mammal_sequences
}
\keyword{datasets}
