% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corlim_hits.R
\name{corlim_hits}
\alias{corlim_hits}
\title{get the hits for a range of correlation limit values}
\usage{
corlim_hits(cd, corlim = seq(0, 1, 0.05), laglim = 0.6)
}
\arguments{
\item{cd}{list of outputs from the ms_fit function for a set of peptides}

\item{corlim}{a sequence of correlation limit values for thresholding the data}

\item{laglim}{maximum acceptable value of lag for each correlation}
}
\description{
get the hits for a range of correlation limit values
}
\keyword{bruker}
