% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corlim_plot.R
\name{corlim_plot}
\alias{corlim_plot}
\title{get the hits for a range of correlation limit values}
\usage{
corlim_plot(psample, pnpep, pcld, presult, pnames, manname = "unknown",
  pcols = c("green", "red", "blue", "black"), warn = T)
}
\arguments{
\item{cd}{list of outputs from the ms_fit function for a set of peptides}

\item{corlim}{a sequence of correlation limit values for thresholding the data}

\item{laglim}{maximum acceptable value of lag for each correlation}
}
\description{
get the hits for a range of correlation limit values
}
\keyword{bruker}
