% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_plot.R
\name{lag_plot}
\alias{lag_plot}
\title{Plot the lag correction and fit a line to it for a peptide sample}
\usage{
lag_plot(markers, hits, no = 1, ylim = c(-1.4, 1), mypoly = 2,
  title = "Lag Plot", corlim = 0.1, laglim = 0.5)
}
\arguments{
\item{fn}{full path of the raw gpm text fi}
}
\description{
Plot the lag correction and fit a line to it for a peptide sample
}
\examples{
lag_plot(markers,hits)
}

