% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpm_load.R
\name{load.gpm.raw}
\alias{load.gpm.raw}
\title{Load a set of peptides from the Global Protein Machine}
\usage{
load.gpm.raw(fn = NA, colno = NA)
}
\arguments{
\item{fn}{full path of the raw gpm text file}
}
\description{
Load a set of peptides from the Global Protein Machine
}

