% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpm_load.R
\name{load.gpm}
\alias{load.gpm}
\title{Load a set of peptides from the Global Protein Machine}
\usage{
load.gpm(fn = "~/tmp/bioarch_keri/161208meeting/gmp_human_collagen.dat")
}
\arguments{
\item{fn}{full path of the raw gpm text file}
}
\description{
Load a set of peptides from the Global Protein Machine
}

