% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_markers.R
\name{load.human.markers}
\alias{load.human.markers}
\title{Load a set of peptide markers for human}
\usage{
load.human.markers(fn = NA)
}
\arguments{
\item{fn}{full path of the raw gpm text file}
}
\description{
Load a set of peptide markers for human
}
\examples{
hm<-load.human.markers()
}
