% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_load.R
\name{load.mcs}
\alias{load.mcs}
\title{Load peptides from the mammalian collagen sequences googlesheet}
\usage{
load.mcs(spp = "human", sheet = bioarch_mammal_sequences, massmin = 800,
  massmax = 3500, verbose = F)
}
\arguments{
\item{species}{the name of the species whose peptides will be loaded.}
}
\description{
Load peptides from the mammalian collagen sequences googlesheet
}
\examples{
hcs <- load.mcs()
hcs <- load.mcs("goat")
}

