% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_load.R
\name{load.sequence}
\alias{load.sequence}
\title{Load sequences from the mammalian collagen sequences googlesheet}
\usage{
load.sequence(spp = "human", sheet = bioarch_mammal_sequences,
  verbose = F, col3 = F)
}
\arguments{
\item{sheet}{the spreadsheet to load the data from. Must be the same format as bioarch_mammal_sequences}

\item{verbose}{verbose processing with more detail. Useful for debugging}

\item{col3}{whether to load the sequence from column3, or from the remainder of the spreadsheet. For debugging.}

\item{species}{the name of the species whose peptides will be loaded.}
}
\description{
Load sequences from the mammalian collagen sequences googlesheet
}
\examples{
hcs <- load.sequence()
hcs <- load.sequence("goat")
}

