% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_align.R
\name{ms_align}
\alias{ms_align}
\title{Align a theoretical peptide with a mass spec}
\usage{
ms_align(
  ts,
  data,
  txlim,
  gauss = NA,
  normlim = NA,
  cctitle = NA,
  doplot = F,
  verbose = F,
  ccylim = c(-0.1, 0.5)
)
}
\arguments{
\item{ts}{theoretical spectrum peaks for a particular peptide}

\item{data}{the MALDI}

\item{txlim}{the range of masses over which to carry out the alignment}

\item{gauss}{the level of gaussian smoothing. defaults to NA (no smoothing)}

\item{normlim}{the minimum upper value that the intensities should be normalised to. Defaults to NA (no minimum)}

\item{doplot}{whether to generate a plot of the alignment}

\item{verbose}{whether to write messages whilst processing}

\item{ccylim}{range of y axis in cross-correlation plot (defaults to [-0.1,0.5])}
}
\value{
a dataframe holding the following fields:
  \item{lag1}{The lag for sample 1}
  \item{lag2}{The lag for sample 2}
  \item{lag3}{The lag for sample 3}
  \item{cor1}{The correlation coefficient for sample 1}
  \item{cor2}{The correlation coefficient for sample 2}
  \item{cor3}{The correlation coefficient for sample 3}
  \item{ion1}{The proportion of total ions for this peptide in sample 1}
  \item{ion2}{The proportion of total ions for this peptide in sample 2}
  \item{ion3}{The proportion of total ions for this peptide in sample 3}
}
\description{
Align a theoretical peptide with a mass spec
}
