% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fit.R
\name{ms_fit}
\alias{ms_fit}
\title{Match a set of peptides with a Mass Spec}
\usage{
ms_fit(peptides, sample)
}
\arguments{
\item{peptides}{the set of peptides}

\item{sample}{the MS sample data}
}
\description{
Match a set of peptides with a Mass Spec
}

