% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_iso.R
\name{ms_iso}
\alias{ms_iso}
\title{Obtain the isotope distribution for an amino acid sequence with
post-translation modifications (deamitation and hydroxylation)}
\usage{
ms_iso(sequence, ndeamidations = 0, nhydroxylations = 0, verbose = F)
}
\arguments{
\item{ndeamidations}{the number of deamidations, default is zero}

\item{nhydroxylations}{the number of hydroxylations, default is zero}

\item{seqeunce}{the amino acid sequence}
}
\description{
Obtain the isotope distribution for an amino acid sequence with
post-translation modifications (deamitation and hydroxylation)
}
\examples{
ms_iso("IGQPGAVGPAGIR")
ms_iso("IGQPGAVGPAGIR",0,1)
}
\keyword{deamidation}
\keyword{hyroxylation}
\keyword{isotopes}
\keyword{mass}
\keyword{sequence}
