% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_iso_atom.R
\name{ms_iso_atom}
\alias{ms_iso_atom}
\title{Obtain the isotope distribution for a molecule consisting of atoms of C,H,N,O and S
post-translation modifications (deamitation and hydroxylation)}
\usage{
ms_iso_atom(atoms, verbose = F)
}
\arguments{
\item{atoms}{vector containing the counts of atoms in the order C,H,N,O and S}
}
\description{
Obtain the isotope distribution for a molecule consisting of atoms of C,H,N,O and S
post-translation modifications (deamitation and hydroxylation)
}
\examples{
ms_iso_atom(c(47,69,15,12,0))
}
\keyword{deamidation}
\keyword{hyroxylation}
\keyword{isotopes}
\keyword{mass}
\keyword{sequence}
