% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_subrange.R
\name{ms_subrange}
\alias{ms_subrange}
\title{Get all spectrum data between a mass range}
\usage{
ms_subrange(ms, lbl, ubl)
}
\arguments{
\item{ms}{the mass spectrum}

\item{lbl}{the lower bound limit}

\item{ubl}{the upper bound limit}
}
\value{
the subrage
}
\description{
Get all spectrum data between a mass range
}
