% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_tpeaks.R
\name{ms_tpeaks}
\alias{ms_tpeaks}
\title{Wrapper for the 'R_iso_seq' q2e c function, to get theoretical peaks from a sequence}
\usage{
ms_tpeaks(sequence)
}
\arguments{
\item{seqeunce}{the amino acid sequence}
}
\description{
Wrapper for the 'R_iso_seq' q2e c function, to get theoretical peaks from a sequence
}
\examples{
ms_tpeaks("IGQPGAVGPAGIR")
}
\keyword{test}

