% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msms_align.R
\name{msms_align}
\alias{msms_align}
\title{Locally align a pair of  mass specs}
\usage{
msms_align(
  data1,
  data2,
  mass,
  alim = c(-1, 6),
  laglim = 0.3,
  by = 0.005,
  doplot = T,
  verbose = T
)
}
\arguments{
\item{data1}{the first MS}

\item{data2}{the secon MS}

\item{mass}{the mass around which to do the alighment}

\item{alim}{the window around the mass, defaluts to [-1,6]}

\item{laglim}{the largest permissable lag}

\item{by}{the resampling rate, default value is 0.005}

\item{doplot}{whether to generate a plot of the alignment}

\item{verbose}{whether to write messages whilst processing}
}
\value{
a dataframe holding the following fields:
  \item{lag}{The lag for sample 1}
  \item{cor}{The correlation coefficient for sample 1}
}
\description{
Locally align a pair of  mass specs
}
