% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pepsheep}
\alias{pepsheep}
\title{Descriminatory peptide markers for sheep}
\format{A data frame with 5 rows and 7 variables:
\describe{
  \item{seq}{The peptide sequence}
  \item{nhyd}{The number of hydroxylations}
  \item{nglut}{The number of deamidations of glutamine}
  \item{mass1}{The mass of the first isotope of the peptide}
  \item{seqpos}{The position in the collagen sequence}
  \item{missed.cleaves}{The number of missed cleaves in the peptide}
  \item{collagen}{The collagen number that the sequence came from}
  \item{prob}{Estimate of the probability of the nhyd value}
}}
\usage{
pepsheep
}
\description{
Descriminatory peptide markers for sheep
}
\keyword{datasets}
