% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pplate.R
\name{pplate}
\alias{pplate}
\title{Process a MALDI plate of replicates}
\usage{
pplate(path, pd, FUN = NULL, ...)
}
\arguments{
\item{path}{the relative path to the data from where the function is called}

\item{pd}{dataframe containing the following fields:
* {froot}{The common part of the file name, e.g. "20211017_Filey_" as part of "20211017_Filey_A1.txt"}
* {sampleID}{A name for the sample, e.g. "A420"}
* {spot1}{The location of spot1, e.g. "A1"}
* {spot2}{The location of spot2, e.g. "A4"}
* {spot3}{The location of spot3, e.g. "A7"}}

\item{FUN}{the processing function, e.g. "averageMSD"}

\item{ts}{theoretical spectrum peaks for a particular peptide}
}
\description{
Process a MALDI plate of replicates
}
