% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_to_atoms.R
\name{seq_to_atoms}
\alias{seq_to_atoms}
\title{Calculate the atoms in an amino acid sequence
post-translation modifications (deamidation and hydroxylation)}
\usage{
seq_to_atoms(sequence, ndeamidations = 0, nhydroxylations = 0,
  verbose = F)
}
\arguments{
\item{ndeamidations}{the number of deamidations}

\item{nhydroxylations}{the number of hydroxylations}

\item{seqeunce}{the amino acid sequence}
}
\value{
integer vector of atoms in the order C H N O S
}
\description{
Calculate the atoms in an amino acid sequence
post-translation modifications (deamidation and hydroxylation)
}
\examples{
seq_to_atoms("IGQPGAVGPAGIR")
}
\keyword{deamidation}
\keyword{hyroxylation}
\keyword{isotopes}
\keyword{mass}
\keyword{sequence}
