% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppdiffs.R
\name{spp.diffs}
\alias{spp.diffs}
\title{get the differences in sequence between two species.
NB the sequences are sorted by sequence position *within* this function,
so there's no need to sort them before calling it.}
\usage{
spp.diffs(mcs1, mcs2, name1 = "spp01", name2 = "spp02", verbose = F, halt = F)
}
\arguments{
\item{mcs1}{the mammalian collagen sequence for species 1}

\item{mcs2}{the mammalian collagen sequence for species 2}

\item{name1}{(optional) the name of species 1}

\item{name2}{(optional) the name of species 2}

\item{verbose}{(optional) whether to print messages while processing}

\item{halt}{(optional) whether to pause until <return> is hit while processing}
}
\value{
Returns a vector of sequence positions for the position where the differences in the spectrum are
}
\description{
get the differences in sequence between two species.
NB the sequences are sorted by sequence position *within* this function,
so there's no need to sort them before calling it.
}
\examples{
sh <- load.mcs("sheep")
hu <- load.mcs()
diffs <- spp.diffs(sh,hu,"sheep","human")
}
