% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_load.R
\name{ts_index}
\alias{ts_index}
\title{Get the species index in the mammalian_collagen_sequences sheet}
\usage{
ts_index(sheet, spp, verbose = F)
}
\arguments{
\item{sheet}{the google sheet to search}

\item{spp}{the name of the species whose peptides will be loaded.}

\item{verbose}{verbose output while running}
}
\description{
Get the species index in the mammalian_collagen_sequences sheet
}
