% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validseq.R
\name{validseq}
\alias{validseq}
\title{This checks if the letters in a sequence are all valid amino acid codes}
\usage{
validseq(seq, halt = F)
}
\arguments{
\item{seq}{character string of an amino acid sequence}

\item{halt}{(optional) whether to pause until <return> is hit while processing}
}
\value{
a logical value: T if valid, F if not
}
\description{
This checks if the letters in a sequence are all valid amino acid codes
}

