% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_pos_prolines.R
\name{xy.ypos.prolines}
\alias{xy.ypos.prolines}
\title{count the prolines in the x and y position of a collagen peptide}
\usage{
xy.ypos.prolines(seq, verbose = F)
}
\arguments{
\item{seq}{the amino acid sequence}

\item{verbose}{default is False}
}
\description{
count the prolines in the x and y position of a collagen peptide
}
\examples{
xy.pos.prolines("IGQPGAVGPAGIR")
}
\keyword{deamidation}
\keyword{hyroxylation}
\keyword{isotopes}
\keyword{mass}
\keyword{sequence}
