% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpm_load.R
\name{load.gpm}
\alias{load.gpm}
\title{Load a set of peptides from the Global Protein Machine}
\usage{
load.gpm(col1fn = NA, col2fn = NA)
}
\arguments{
\item{col1fn}{full path of the raw gpm text file for collagen 1}

\item{col2fn}{full path of the raw gpm text file for collagen 1}
}
\description{
Load a set of peptides from the Global Protein Machine
}
\examples{
gpm<-load.gpm()
}
