% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_load.R
\name{load.sample}
\alias{load.sample}
\title{Load mass spec data from a 2-column space-delimited text file}
\usage{
load.sample(froot, name = "Sample", spots)
}
\arguments{
\item{species}{the name of the species whose peptides will be loaded.}
}
\description{
Load mass spec data from a 2-column space-delimited text file
}

