% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fit.R
\name{ms_fit}
\alias{ms_fit}
\title{Match a set of peptides with a Mass Spec}
\usage{
ms_fit(peptides, sample, doplot = T, forceplot = F, vlevel = 0,
  corlim = 0, laglim = 0.6)
}
\arguments{
\item{peptides}{the set of peptides}

\item{sample}{the MS sample data}
}
\value{
a logical vector stating which peptides match the data
}
\description{
Match a set of peptides with a Mass Spec
}

