% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_tpeaks.R
\name{ms_tpeaks}
\alias{ms_tpeaks}
\title{Wrapper for the 'R_iso_seq' q2e c function, to get theoretical peaks from a sequence
WARNING: This function is obsolete - use ms_fit instead with the same arguments}
\usage{
ms_tpeaks(sequence, verbose = F, ndeamidations = 0, nhydroxylations = 0)
}
\arguments{
\item{seqeunce}{the amino acid sequence}
}
\description{
Wrapper for the 'R_iso_seq' q2e c function, to get theoretical peaks from a sequence
WARNING: This function is obsolete - use ms_fit instead with the same arguments
}
\examples{
ms_tpeaks("IGQPGAVGPAGIR")
}
\keyword{test}
