% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_seq.R
\name{parse.seq}
\alias{parse.seq}
\title{Parse an amino acid sequence into a set of peptides, incorporating hydroxylation and deamidation modifications, and calculating mass}
\usage{
parse.seq(
  sequence,
  cuts = "K|R",
  skip = "X",
  massmin = 800,
  massmax = 3500,
  verbose = F,
  max.missed.cleaves = 0,
  cutbefore = F
)
}
\arguments{
\item{cuts}{a regular expression listing the amino acids that the sequence should be cut at. Defaults to Trypsing cut sites K & R}

\item{skip}{any codes that mean the peptide is invalid. Defaults to "X"}

\item{massmin}{the minimum mass that a peptide should be. Defaults to 800}

\item{massmax}{the maximum mass that a peptide should be. Defaults to 3500}

\item{verbose}{whether to print messages to console during processing. Defaults to FALSE}

\item{max.missed.cleaves}{the number of missed cleaves to incorporate in the set of peptides.}

\item{seqeunce}{a character string holding the sequence, e.g. "GPPGAPGPPGPP"}
}
\description{
Parse an amino acid sequence into a set of peptides, incorporating hydroxylation and deamidation modifications, and calculating mass
}
\examples{
parse.seq("GPPGQKGPPGPQGPRGPPGPPGPM")
}
