% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratios.R
\name{peptide.ratio}
\alias{peptide.ratio}
\title{Get the ratio of ion counts for two peptides}
\usage{
peptide.ratio(sample, peptide1, peptide2, doplot = F)
}
\arguments{
\item{sample}{the MALDI spectrum - three replictes, as loaded by load.sample()}

\item{peptide1}{the first peptide, which forms the numerator of the ratio calculation}

\item{peptide2}{the second peptide, which forms the denominator of the ratio calculation}

\item{doplot}{whether or not a plot of the fit should be created (default is FALSE)}
}
\description{
Get the ratio of ion counts for two peptides
}
\examples{
sample <- load.sample("~/samples","01",c("_1","_2","_3"))
peptides <- load.mcs()
ratios <- peptide.ratio(sample, peptides[1,], peptides[2,])
}
